ai-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
google-generativeai==0.8.3
pydantic==2.9.2
python-dotenv==1.0.1
pandas==2.2.3
numpy==2.1.3
pymongo==4.9.1
motor==3.6.0
python-multipart==0.0.12
httpx==0.27.2
